requirements.txt
streamlit
requests
python-dotenv
pypdf2